factor out DefaultValueSource
